src/requirements.txt
appdirs==1.4.3
certifi==2018.10.15
chardet==3.0.4
Django==1.11.27
django-bootstrap3==11.0.0
django-ckeditor==5.6.1
django-debug-toolbar==1.7
django-js-asset==1.1.0
django-recaptcha2==1.3.1
django-tagging==0.4.6
django-taggit==0.23.0
idna==2.7
olefile==0.46
packaging==16.8
Pillow==5.3.0
psycopg2==2.7.5
pyparsing==2.2.2
pytz==2018.7
requests==2.20.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24
uWSGI==2.0.17.1
django-environ==0.4.5
